adjust

+++ src/app/modules/customer/component/region/region.component.ts
@@ -2,11 +2,15 @@ import { Component, OnInit } from '@angular/core';
 import { Region } from '../../_model/region';
 import { RegionService } from '../../_service/region.service';
 import { FormBuilder, Validators} from '@angular/forms';
+
 declare var $: any;
+
+import Swal from 'sweetalert2';
+
 @Component({
   selector: 'app-region',
   templateUrl: './region.component.html',
-  styleUrls: ['./region.component.scss']
+  styleUrls: ['./region.component.css']
 })
 export class RegionComponent implements OnInit {
 
@@ -56,9 +60,22 @@ export class RegionComponent implements OnInit {
           console.log(this.region);
           this.getRegions();
           this.closeModal();
-
+          Swal.fire({
+            position: 'top-end',
+            icon: 'success',
+            title: 'Registro exitoso!',
+            showConfirmButton: false,
+            timer: 1500
+          })
         },
-        err => console.log(err)
+        err => {
+          console.log(err);
+          Swal.fire({
+            icon: 'error',
+            title: 'Error!',
+            text: 'La región no puede ser registrada',
+          })
+        }
       )
     }else{
       this.region_service.updateRegion(this.formulario.value).subscribe(
@@ -66,10 +83,22 @@ export class RegionComponent implements OnInit {
           console.log(this.region);
           this.getRegions();
           this.closeModal();
-          
-
+          Swal.fire({
+            position: 'top-end',
+            icon: 'success',
+            title: 'Actualización exitosa!',
+            showConfirmButton: false,
+            timer: 1500
+          })
         },
-        err => console.log(err)
+        err => {
+          console.log(err);
+          Swal.fire({
+            icon: 'error',
+            title: 'Error!',
+            text: 'La región no puede ser actualizada',
+          })
+        }
       )
     }
   }
@@ -77,8 +106,6 @@ export class RegionComponent implements OnInit {
   createRegion(){
     this.post_region = true;
     this.formulario.reset();
-    $('#modal-tittle').text("Crear Región");
-    $('#modal-submit').text("Crear");
     $("#region_modal").modal("show");
   }
 
@@ -86,19 +113,43 @@ export class RegionComponent implements OnInit {
     this.post_region = false;
     this.formulario.controls['id_region'].setValue(region.id_region);
     this.formulario.controls['region'].setValue(region.region);
-    $('#modal-tittle').text("Actualizar Región");
-    $('#modal-submit').text("Actualizar");
     $("#region_modal").modal("show");
   }
 
   deleteRegion(id_region: number){
-    this.region_service.deleteRegion(id_region).subscribe(
-      res => {
-        console.log(this.region);
-        this.getRegions();
-      },
-      err => console.log(err)
-    )
+    Swal.fire({
+      title: 'Deseas eliminar la Región?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si, eliminar!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.region_service.deleteRegion(id_region).subscribe(
+          res => {
+            console.log(this.region);
+            this.getRegions();
+            Swal.fire({
+              position: 'top-end',
+              icon: 'success',
+              title: 'Eliminación exitosa!',
+              showConfirmButton: false,
+              timer: 1500
+            })
+          },
+          err => {
+            console.log(err);
+            Swal.fire({
+              icon: 'error',
+              title: 'Error!',
+              text: 'La región no puede ser eliminada',
+            })
+          }
+        )
+      }
+    })
+    
   }
 
   get f() {
